% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhteds_helper_functions.R
\name{bhteds_grid_table_theme}
\alias{bhteds_grid_table_theme}
\title{BHTEDS Project Data Visualizations Helper Function}
\usage{
bhteds_grid_table_theme(table = NULL)
}
\arguments{
\item{table}{Dataframe containing a returned aggregate statistics table from the visualize_demographics() function.}

\item{...}{To be implemented.}
}
\value{
gridExtra::ttheme_default() object.
}
\description{
Generates a modified theme for the tableGrob() object that visualizes BHTEDS aggregate statistics.
}
\details{
Assumes 7 columns in the aggregate statistics dataframe/table.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{visualize_demographics}}
}
\author{
Joseph Jinn, \email{jinnjo@mail.gvsu.edu}
}
